bevy_ecs_ldtk
Module app

Types and traits for hooking into the ldtk loading process via bevy::app::App.

Structs

PhantomLdtkEntity

PhantomLdtkIntCell

Traits

LdtkEntity

Provides a constructor which can be used for spawning entities from an LDtk file.

LdtkEntityAppExt

Provides functions to register Bundles to bevy’s App for particular LDtk layer and entity identifiers.

LdtkIntCell

Provides a constructor which can be used for spawning additional components on IntGrid tiles.

LdtkIntCellAppExt

Provides functions to register Bundles to bevy’s App for particular LDtk layer and IntGrid values.

PhantomLdtkEntityTrait

PhantomLdtkIntCellTrait

Type Aliases

LdtkEntityMap

Used by LdtkEntityAppExt to associate Ldtk entity identifiers with LdtkEntitys.

LdtkIntCellMap

Used by LdtkIntCellAppExt to associate Ldtk IntGrid values with LdtkIntCells.