bevy_ecs_ldtk::app
Trait LdtkIntCellAppExt

pub trait LdtkIntCellAppExt {
    // Required method
    fn register_ldtk_int_cell_for_layer_optional<B: LdtkIntCell + Bundle>(
        &mut self,
        layer_identifier: Option<String>,
        value: Option<i32>,
    ) -> &mut Self;

    // Provided methods
    fn register_ldtk_int_cell_for_layer<B: LdtkIntCell + Bundle>(
        &mut self,
        layer_identifier: &str,
        value: i32,
    ) -> &mut Self { ... }
    fn register_ldtk_int_cell<B: LdtkIntCell + Bundle>(
        &mut self,
        value: i32,
    ) -> &mut Self { ... }
    fn register_default_ldtk_int_cell_for_layer<B: LdtkIntCell + Bundle>(
        &mut self,
        layer_identifier: &str,
    ) -> &mut Self { ... }
    fn register_default_ldtk_int_cell<B: LdtkIntCell + Bundle>(
        &mut self,
    ) -> &mut Self { ... }
}

Provides functions to register Bundles to bevy’s App for particular LDtk layer and IntGrid values.

After being registered, entities will be spawned with these Bundles when some IntGrid cell in LDtk meets the criteria you specify.

Not intended for custom implementations on your own types.

Required Methods

fn register_ldtk_int_cell_for_layer_optional<B: LdtkIntCell + Bundle>( &mut self, layer_identifier: Option<String>, value: Option<i32>, ) -> &mut Self

Used internally by all the other LDtk int cell registration functions.

Similar to LdtkIntCellAppExt::register_ldtk_int_cell_for_layer, except it provides defaulting functionality:

Setting layer_identifier to None will make the registration apply to any IntGrid layer.

Setting value to None will make the registration apply to any IntGrid tile.

This defaulting functionality means that a particular LDtk IntGrid tile may match multiple registrations. In these cases, registrations are prioritized in order of most to least specific:

layer_identifier and value are specified

Just value is specified

Just layer_identifier is specified

Neither value nor layer_identifier are specified

Provided Methods

fn register_ldtk_int_cell_for_layer<B: LdtkIntCell + Bundle>( &mut self, layer_identifier: &str, value: i32, ) -> &mut Self

Registers LdtkIntCell types to be inserted for a given IntGrid value and layer identifier in an LDtk file.

This example lets the plugin know that it should spawn a MyBundle when it encounters an IntGrid tile whose value is 1 on a “MyLayerIdentifier” layer.

use bevy::prelude::*;
use bevy_ecs_ldtk::prelude::*;

fn main() {
    App::empty()
        .add_plugins(LdtkPlugin)
        .register_ldtk_int_cell_for_layer::<MyBundle>("MyLayerIdentifier", 1)
        // add other systems, plugins, resources...
        .run();
}

#[derive(Bundle, LdtkIntCell, Default)]
pub struct MyBundle {
    a: ComponentA,
    b: ComponentB,
    c: ComponentC,
}

You can find more details on the #[derive(LdtkIntCell)] macro at LdtkIntCell.

fn register_ldtk_int_cell<B: LdtkIntCell + Bundle>( &mut self, value: i32, ) -> &mut Self

Similar to LdtkIntCellAppExt::register_ldtk_int_cell_for_layer, except it applies the registration to all layers.

fn register_default_ldtk_int_cell_for_layer<B: LdtkIntCell + Bundle>( &mut self, layer_identifier: &str, ) -> &mut Self

Similar to LdtkIntCellAppExt::register_ldtk_int_cell_for_layer, except it applies the registration to all tiles on the given layer.

fn register_default_ldtk_int_cell<B: LdtkIntCell + Bundle>( &mut self, ) -> &mut Self

Similar to LdtkIntCellAppExt::register_ldtk_int_cell_for_layer, except it applies the registration to any tile and any layer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.