Module bevy_ecs_ldtk::assets

Assets and related items for loading LDtk files.

Structs

InternalLevels

Marker type for indicating an internal-levels LDtk project.

LdtkAssetPlugin

Plugin that registers LDtk-related assets.

LdtkJsonWithMetadata

LDtk json data and level metadata produced when loading an LdtkProject asset.

LdtkProject

Main asset for loading LDtk project data.

LevelIndices

Indices pointing to the location of a level in an LdtkProject or LdtkJson.

LevelMetadata

Metadata produced for every level during LdtkProject loading.

Enums

LdtkProjectData

LDtk json data and level metadata for both internal- and external-level projects.

Traits

LevelMetadataAccessor

Convenience methods for types that store levels and level metadata.