bevy_ecs_ldtk::assets
Enum LdtkProjectData

pub enum LdtkProjectData {
    Standalone(LdtkJsonWithMetadata<InternalLevels>),
}

LDtk json data and level metadata for both internal- and external-level projects.

We need to abstract over these cases to allow them in the same asset type: LdtkProject. All methods that are available in both cases are available here. However, methods exclusive to each case require accessing the internal type. These include methods for obtaining LoadedLevels. See the LoadedLevel-accessing methods in the following impls:

standalone projects

parent projects

Note that this type’s variants are under different feature flags. At least one of these feature flags needs to be enabled for the plugin to compile.

Variants

Standalone(LdtkJsonWithMetadata<InternalLevels>)

LDtk data for a standalone project (uses internal levels).

This is only available under the internal_levels feature.

Implementations

impl LdtkProjectData

pub fn json_data(&self) -> &LdtkJson

Raw ldtk json data.

pub fn as_standalone(&self) -> &LdtkJsonWithMetadata<InternalLevels>

Unwrap as a LdtkJsonWithMetadata<InternalLevels>. For use on internal-levels ldtk projects only.

Panics

Panics if this is not LdtkProjectData::Standalone. This shouldn’t occur if the project uses internal levels.