Struct bevy_ecs_ldtk::assets::LdtkJsonWithMetadata

pub struct LdtkJsonWithMetadata<L>
where
    L: LevelLocale,
    L::Metadata: TypePath,
{ /* private fields */ }

LDtk json data and level metadata produced when loading an LdtkProject asset.

Generic over a level-locale marker type, L. This helps differentiate between internal- and external-level projects. L will can only be either InternalLevels or [ExternalLevels]. This provides some abstraction over the two cases, but they are ultimately different types. Some methods are exclusive to each case, especially for obtaining LoadedLevels. See the LoadedLevel-accessing methods in the following impls:

internal-levels

external-levels

Implementations

impl<L> LdtkJsonWithMetadata<L>
where L: LevelLocale, L::Metadata: TypePath,

Auto-generated by derive_getters::Getters.

pub fn json_data(&self) -> &LdtkJson

Raw ldtk json data.

pub fn level_map(&self) -> &HashMap<String, L::Metadata>

Map from level iids to level metadata.

impl LdtkJsonWithMetadata<InternalLevels>

pub fn iter_loaded_levels(&self) -> impl Iterator<Item = LoadedLevel<'_>>

Iterate through this project’s loaded levels.

This first iterates through root levels, then world levels.

These levels are loaded, meaning that they are type-guaranteed to have complete data.

pub fn get_loaded_level_at_indices( &self, indices: &LevelIndices ) -> Option<LoadedLevel<'_>>

Immutable access to a loaded level at the given LevelIndices.

These levels are loaded, meaning that they are type-guaranteed to have complete data.

pub fn get_loaded_level_by_iid(&self, iid: &String) -> Option<LoadedLevel<'_>>

Returns a reference to the loaded level metadata corresponding to the given level iid.

These levels are loaded, meaning that they are type-guaranteed to have complete data.

pub fn find_loaded_level_by_level_selection( &self, level_selection: &LevelSelection ) -> Option<LoadedLevel<'_>>

Find the loaded level matching the given LevelSelection.

This lookup is constant for LevelSelection::Iid and LevelSelection::Indices variants. The other variants require iterating through the levels to find the match.

These levels are loaded, meaning that they are type-guaranteed to have complete data.