bevy_ecs_ldtk::ldtk::ldtk_fields
Trait LdtkFields

pub trait LdtkFields {

    // Required method
    fn field_instances(&self) -> &[FieldInstance];

    // Provided methods
    fn get_field_instance(
        &self,
        identifier: &str,
    ) -> Result<&FieldInstance, LdtkFieldsError> { ... }
    fn get_field(
        &self,
        identifier: &str,
    ) -> Result<&FieldValue, LdtkFieldsError> { ... }
    fn get_maybe_int_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<i32>, LdtkFieldsError> { ... }
    fn get_int_field(&self, identifier: &str) -> Result<&i32, LdtkFieldsError> { ... }
    fn get_maybe_float_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<f32>, LdtkFieldsError> { ... }
    fn get_float_field(&self, identifier: &str) -> Result<&f32, LdtkFieldsError> { ... }
    fn get_bool_field(&self, identifier: &str) -> Result<&bool, LdtkFieldsError> { ... }
    fn get_maybe_string_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<String>, LdtkFieldsError> { ... }
    fn get_string_field(
        &self,
        identifier: &str,
    ) -> Result<&String, LdtkFieldsError> { ... }
    fn get_color_field(
        &self,
        identifier: &str,
    ) -> Result<&Color, LdtkFieldsError> { ... }
    fn get_maybe_file_path_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<String>, LdtkFieldsError> { ... }
    fn get_file_path_field(
        &self,
        identifier: &str,
    ) -> Result<&String, LdtkFieldsError> { ... }
    fn get_maybe_enum_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<String>, LdtkFieldsError> { ... }
    fn get_enum_field(
        &self,
        identifier: &str,
    ) -> Result<&String, LdtkFieldsError> { ... }
    fn get_maybe_tile_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<TilesetRectangle>, LdtkFieldsError> { ... }
    fn get_tile_field(
        &self,
        identifier: &str,
    ) -> Result<&TilesetRectangle, LdtkFieldsError> { ... }
    fn get_maybe_entity_ref_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<ReferenceToAnEntityInstance>, LdtkFieldsError> { ... }
    fn get_entity_ref_field(
        &self,
        identifier: &str,
    ) -> Result<&ReferenceToAnEntityInstance, LdtkFieldsError> { ... }
    fn get_maybe_point_field(
        &self,
        identifier: &str,
    ) -> Result<&Option<IVec2>, LdtkFieldsError> { ... }
    fn get_point_field(
        &self,
        identifier: &str,
    ) -> Result<&IVec2, LdtkFieldsError> { ... }
    fn get_maybe_ints_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<i32>], LdtkFieldsError> { ... }
    fn iter_ints_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, i32>, LdtkFieldsError> { ... }
    fn get_maybe_floats_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<f32>], LdtkFieldsError> { ... }
    fn iter_floats_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, f32>, LdtkFieldsError> { ... }
    fn get_bools_field(
        &self,
        identifier: &str,
    ) -> Result<&[bool], LdtkFieldsError> { ... }
    fn get_maybe_strings_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<String>], LdtkFieldsError> { ... }
    fn iter_strings_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, String>, LdtkFieldsError> { ... }
    fn get_colors_field(
        &self,
        identifier: &str,
    ) -> Result<&[Color], LdtkFieldsError> { ... }
    fn get_maybe_file_paths_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<String>], LdtkFieldsError> { ... }
    fn iter_file_paths_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, String>, LdtkFieldsError> { ... }
    fn get_maybe_enums_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<String>], LdtkFieldsError> { ... }
    fn iter_enums_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, String>, LdtkFieldsError> { ... }
    fn get_maybe_tiles_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<TilesetRectangle>], LdtkFieldsError> { ... }
    fn iter_tiles_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, TilesetRectangle>, LdtkFieldsError> { ... }
    fn get_maybe_entity_refs_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<ReferenceToAnEntityInstance>], LdtkFieldsError> { ... }
    fn iter_entity_refs_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, ReferenceToAnEntityInstance>, LdtkFieldsError> { ... }
    fn get_maybe_points_field(
        &self,
        identifier: &str,
    ) -> Result<&[Option<IVec2>], LdtkFieldsError> { ... }
    fn iter_points_field(
        &self,
        identifier: &str,
    ) -> Result<AllSomeIter<'_, IVec2>, LdtkFieldsError> { ... }

}

Convenience methods for accessing field instances.

Required Methods

fn field_instances(&self) -> &[FieldInstance]

Immutable accessor for this item’s field instances, by reference.

Provided Methods

fn get_field_instance( &self, identifier: &str, ) -> Result<&FieldInstance, LdtkFieldsError>

Get this item’s field instance (with metadata) for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

fn get_field(&self, identifier: &str) -> Result<&FieldValue, LdtkFieldsError>

Get this item’s field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

fn get_maybe_int_field( &self, identifier: &str, ) -> Result<&Option<i32>, LdtkFieldsError>

Get this item’s nullable Int field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Int.

fn get_int_field(&self, identifier: &str) -> Result<&i32, LdtkFieldsError>

Get this item’s non-null Int field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Int.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_maybe_float_field( &self, identifier: &str, ) -> Result<&Option<f32>, LdtkFieldsError>

Get this item’s nullable Float field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Float.

fn get_float_field(&self, identifier: &str) -> Result<&f32, LdtkFieldsError>

Get this item’s non-null Float field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Float.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_bool_field(&self, identifier: &str) -> Result<&bool, LdtkFieldsError>

Get this item’s Bool field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Bool.

fn get_maybe_string_field( &self, identifier: &str, ) -> Result<&Option<String>, LdtkFieldsError>

Get this item’s nullable String field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::String.

fn get_string_field(&self, identifier: &str) -> Result<&String, LdtkFieldsError>

Get this item’s non-null String field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::String.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_color_field(&self, identifier: &str) -> Result<&Color, LdtkFieldsError>

Get this item’s Color field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Color.

fn get_maybe_file_path_field( &self, identifier: &str, ) -> Result<&Option<String>, LdtkFieldsError>

Get this item’s nullable FilePath field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::FilePath.

fn get_file_path_field( &self, identifier: &str, ) -> Result<&String, LdtkFieldsError>

Get this item’s non-null FilePath field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::FilePath.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_maybe_enum_field( &self, identifier: &str, ) -> Result<&Option<String>, LdtkFieldsError>

Get this item’s nullable Enum field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Enum.

fn get_enum_field(&self, identifier: &str) -> Result<&String, LdtkFieldsError>

Get this item’s non-null Enum field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Enum.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_maybe_tile_field( &self, identifier: &str, ) -> Result<&Option<TilesetRectangle>, LdtkFieldsError>

Get this item’s nullable Tile field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Tile.

fn get_tile_field( &self, identifier: &str, ) -> Result<&TilesetRectangle, LdtkFieldsError>

Get this item’s non-null Tile field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Tile.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_maybe_entity_ref_field( &self, identifier: &str, ) -> Result<&Option<ReferenceToAnEntityInstance>, LdtkFieldsError>

Get this item’s nullable EntityRef field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::EntityRef.

fn get_entity_ref_field( &self, identifier: &str, ) -> Result<&ReferenceToAnEntityInstance, LdtkFieldsError>

Get this item’s non-null EntityRef field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::EntityRef.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_maybe_point_field( &self, identifier: &str, ) -> Result<&Option<IVec2>, LdtkFieldsError>

Get this item’s nullable Point field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Point.

fn get_point_field(&self, identifier: &str) -> Result<&IVec2, LdtkFieldsError>

Get this item’s non-null Point field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Point.

returns LdtkFieldsError::UnexpectedNull if the field is null.

fn get_maybe_ints_field( &self, identifier: &str, ) -> Result<&[Option<i32>], LdtkFieldsError>

Get this item’s nullable Ints field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Ints.

fn iter_ints_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, i32>, LdtkFieldsError>

Iterate over this item’s non-null Ints field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Ints.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_maybe_floats_field( &self, identifier: &str, ) -> Result<&[Option<f32>], LdtkFieldsError>

Get this item’s nullable Floats field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Floats.

fn iter_floats_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, f32>, LdtkFieldsError>

Iterate over this item’s non-null Floats field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Floats.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_bools_field(&self, identifier: &str) -> Result<&[bool], LdtkFieldsError>

Get this item’s Bools field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Bools.

fn get_maybe_strings_field( &self, identifier: &str, ) -> Result<&[Option<String>], LdtkFieldsError>

Get this item’s nullable Strings field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Strings.

fn iter_strings_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, String>, LdtkFieldsError>

Iterate over this item’s non-null Strings field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Strings.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_colors_field( &self, identifier: &str, ) -> Result<&[Color], LdtkFieldsError>

Get this item’s Colors field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Colors.

fn get_maybe_file_paths_field( &self, identifier: &str, ) -> Result<&[Option<String>], LdtkFieldsError>

Get this item’s nullable FilePaths field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::FilePaths.

fn iter_file_paths_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, String>, LdtkFieldsError>

Iterate over this item’s non-null FilePaths field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::FilePaths.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_maybe_enums_field( &self, identifier: &str, ) -> Result<&[Option<String>], LdtkFieldsError>

Get this item’s nullable Enums field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Enums.

fn iter_enums_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, String>, LdtkFieldsError>

Iterate over this item’s non-null Enums field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Enums.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_maybe_tiles_field( &self, identifier: &str, ) -> Result<&[Option<TilesetRectangle>], LdtkFieldsError>

Get this item’s nullable Tiles field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Tiles.

fn iter_tiles_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, TilesetRectangle>, LdtkFieldsError>

Iterate over this item’s non-null Tiles field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Tiles.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_maybe_entity_refs_field( &self, identifier: &str, ) -> Result<&[Option<ReferenceToAnEntityInstance>], LdtkFieldsError>

Get this item’s nullable EntityRefs field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::EntityRefs.

fn iter_entity_refs_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, ReferenceToAnEntityInstance>, LdtkFieldsError>

Iterate over this item’s non-null EntityRefs field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::EntityRefs.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.

fn get_maybe_points_field( &self, identifier: &str, ) -> Result<&[Option<IVec2>], LdtkFieldsError>

Get this item’s nullable Points field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Points.

fn iter_points_field( &self, identifier: &str, ) -> Result<AllSomeIter<'_, IVec2>, LdtkFieldsError>

Iterate over this item’s non-null Points field value for the given identifier.

Errors

returns LdtkFieldsError::FieldNotFound if no field with the given identifier exists.

returns LdtkFieldsError::WrongFieldType if the field is not FieldValue::Points.

returns LdtkFieldsError::UnexpectedNull if any element of the field is null.