Struct bevy_ecs_ldtk::ldtk::LayerInstance

pub struct LayerInstance {

    pub c_hei: i32,
    pub c_wid: i32,
    pub grid_size: i32,
    pub identifier: String,
    pub opacity: f32,
    pub px_total_offset_x: i32,
    pub px_total_offset_y: i32,
    pub tileset_def_uid: Option<i32>,
    pub tileset_rel_path: Option<String>,
    pub layer_instance_type: Type,
    pub auto_layer_tiles: Vec<TileInstance>,
    pub entity_instances: Vec<EntityInstance>,
    pub grid_tiles: Vec<TileInstance>,
    pub iid: String,
    pub int_grid: Option<Vec<IntGridValueInstance>>,
    pub int_grid_csv: Vec<i32>,
    pub layer_def_uid: i32,
    pub level_id: i32,
    pub optional_rules: Vec<i32>,
    pub override_tileset_uid: Option<i32>,
    pub px_offset_x: i32,
    pub px_offset_y: i32,
    pub seed: i32,
    pub visible: bool,

}

Fields
c_hei: i32

Grid-based height

c_wid: i32

Grid-based width

grid_size: i32

Grid size

identifier: String

Layer definition identifier

opacity: f32

Layer opacity as Float [0-1]

px_total_offset_x: i32

Total layer X pixel offset, including both instance and definition offsets.

px_total_offset_y: i32

Total layer Y pixel offset, including both instance and definition offsets.

tileset_def_uid: Option<i32>

The definition UID of corresponding Tileset, if any.

tileset_rel_path: Option<String>

The relative path to corresponding Tileset, if any.

layer_instance_type: Type

Layer type (possible values: IntGrid, Entities, Tiles or AutoLayer)

auto_layer_tiles: Vec<TileInstance>

An array containing all tiles generated by Auto-layer rules. The array is already sorted in display order (ie. 1st tile is beneath 2nd, which is beneath 3rd etc.).

Note: if multiple tiles are stacked in the same cell as the result of different rules, all tiles behind opaque ones will be discarded.

entity_instances: Vec<EntityInstance>grid_tiles: Vec<TileInstance>iid: String

Unique layer instance identifier

int_grid: Option<Vec<IntGridValueInstance>>

WARNING: this deprecated value will be removed completely on version 1.0.0+ Replaced by: intGridCsv

int_grid_csv: Vec<i32>

A list of all values in the IntGrid layer, stored in CSV format (Comma Separated Values).
Order is from left to right, and top to bottom (ie. first row from left to right, followed by second row, etc).
0 means “empty cell” and IntGrid values start at 1.
The array size is __cWid x __cHei cells.

layer_def_uid: i32

Reference the Layer definition UID

level_id: i32

Reference to the UID of the level containing this layer instance

optional_rules: Vec<i32>

An Array containing the UIDs of optional rules that were enabled in this specific layer instance.

override_tileset_uid: Option<i32>

This layer can use another tileset by overriding the tileset UID here.

px_offset_x: i32

X offset in pixels to render this layer, usually 0 (IMPORTANT: this should be added to the LayerDef optional offset, see __pxTotalOffsetX)

px_offset_y: i32

Y offset in pixels to render this layer, usually 0 (IMPORTANT: this should be added to the LayerDef optional offset, see __pxTotalOffsetY)

seed: i32

Random seed used for Auto-Layers rendering

visible: bool

Layer instance visibility