Module bevy_ecs_ldtk::prelude

use bevy_ecs_ldtk::prelude::*; to import commonly used items.

Re-exports

pub use crate::app::LdtkEntity;

pub use crate::app::LdtkEntityAppExt;

pub use crate::app::LdtkIntCell;

pub use crate::app::LdtkIntCellAppExt;

pub use crate::components::EntityInstance;

pub use crate::ldtk;

pub use crate::ldtk::ldtk_fields::LdtkFields;

pub use crate::ldtk::FieldValue;

pub use crate::ldtk::LayerInstance;

pub use crate::ldtk::TilesetDefinition;

Structs

EntityIid

Component added to all LdtkEntitys by default.

GridCoords

Component that stores grid-based coordinate information.

IntGridCell

Component added to any IntGrid tile by default.

LayerMetadata

Component for storing some LDtk layer information on layer entities.

LdtkAsset

Main asset for loading ldtk files.

LdtkLevel

Secondary asset for loading ldtk files, specific to level data.

LdtkPlugin

Adds the default systems, assets, and resources used by bevy_ecs_ldtk.

LdtkSettings

Settings resource for the plugin. Check out the documentation for each field type to learn more.

LdtkWorldBundle

Bundle for spawning LDtk worlds and their levels. The main bundle for using this plugin.

LevelSet

Component that determines the desired levels to be loaded for an LdtkWorldBundle.

ProcessLdtkApi

Schedule for processing this plugin’s ECS API, inserted after Update.

Respawn

Component that indicates that an LDtk level or world should respawn.

TileEnumTags

Component for storing user-defined, enum-based tags for a particular tile in an LDtk tileset definition.

TileMetadata

Component for storing user-defined custom data for a paticular tile in an LDtk tileset definition.

Worldly

Component that indicates that an ldtk entity should be a child of the world, not the level.

Enums

IntGridRendering

Option in LdtkSettings that determines the visual representation of IntGrid layers when they don’t have AutoTile rules.

LevelBackground

Option in LdtkSettings that dictates how the plugin handles level backgrounds.

LevelEvent

Events fired by the plugin related to level spawning/despawning.

LevelSelection

Resource for choosing which level(s) to spawn.

LevelSpawnBehavior

Option in LdtkSettings that determines level spawn behavior.

SetClearColor

Option in LdtkSettings that determines clear color behavior.

Derive Macros

LdtkEntity

LdtkIntCell