Function bevy_ecs_ldtk::utils::ldtk_grid_coords_to_grid_coords

pub fn ldtk_grid_coords_to_grid_coords(
    ldtk_coords: IVec2,
    ldtk_grid_height: i32
) -> GridCoords

Performs LDtk grid coordinate to GridCoords conversion.

This conversion is performed so that both the LDtk grid coords and the resulting GridCoords refer to the same tile. This is different from them referring to the same position in space, because the tile is referenced by its top-left corner in LDtk, and by its bottom-left corner with GridCoords.