Crate bevy_ecs_ldtk

An ECS-friendly LDtk plugin for bevy. Uses bevy_ecs_tilemap as a base.

Getting Started

The goal of this plugin is to make it as easy as possible to use LDtk with bevy for common use cases, while providing solutions to handle more difficult cases. You only need a few things to get started:

Add the LdtkPlugin to the App

Insert the LevelSelection resource into the App to pick your level

Spawn an LdtkWorldBundle

Optionally, use #[derive(LdtkEntity)] and #[derive(LdtkIntCell)] on bundles and register them to the App to automatically spawn those bundles on Entity and IntGrid layers.

use bevy::prelude::*;
use bevy_ecs_ldtk::prelude::*;

fn main() {
    App::new()
        .add_plugins(DefaultPlugins)
        .add_plugins(LdtkPlugin)
        .add_systems(Startup, setup)
        .insert_resource(LevelSelection::Index(0))
        .register_ldtk_entity::<MyBundle>("MyEntityIdentifier")
        .run();
}

fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
    commands.spawn(Camera2dBundle::default());

    commands.spawn(LdtkWorldBundle {
        ldtk_handle: asset_server.load("my_project.ldtk"),
        ..Default::default()
    });
}

#[derive(Bundle, LdtkEntity)]
pub struct MyBundle {
    a: ComponentA,
    b: ComponentB,
    #[sprite_sheet_bundle]
    sprite_bundle: SpriteSheetBundle,
}

Entity and IntGrid layers

You have two options for interacting with Entity and IntGrid LDtk layers.

As mentioned above, you can use #[derive(LdtkEntity)] and #[derive(LdtkIntCell)] on your bundles to hook into the entity/intgrid-tile spawning process. Of course, you can also implement these traits manually. There are some field attribute macros available to these derives to handle the most common use cases. See app::LdtkEntity and app::LdtkIntCell for more details.

You can query for Added<EntityInstance> and Added<IntGridCell> components in a system. This works because, if an LDtk entity or intgrid tile does not match any of your registrations, they are spawned with these components by default. Then, you can use Commands to flesh out these entities like you would normally.

The first option can be convenient and fast, while the second is good if you need more access to the world than the app trait methods provide.

Regardless of your choice, the spawned entities will have an appropriate Transform. They will also be spawned and despawned along with the levels they belong to, unless otherwise specified with a Worldly component. This is because, by default, the entities are spawned as children of the level entities.

Worlds and Levels

When you spawn an LdtkWorldBundle, level entities are automatically spawned as children to the world based off your level selection. The documentation for LdtkWorldBundle goes into a little more detail about the spawning process.

You can select what levels to spawn via the LevelSelection resource, or via the LevelSet component in the LdtkWorldBundle. The LevelSelection resource is a convenient abstraction over the LevelSet component, and updates the LevelSet component automatically when used. It also responds to LevelSpawnBehavior::UseWorldTranslation::load_level_neighbors, while LevelSet does not.

To spawn a new level, you can just update the LevelSelection resource. The current level will be automatically despawned, unless it’s still selected due to loading level neighbors. Updating the LevelSet component will have similar results.

By default, the levels will be spawned so their bottom left corner is at the origin of the world. You can make them spawn according to their world location in LDtk by setting LevelSpawnBehavior::UseWorldTranslation.

Feature flags

This crate uses the following set of feature flags:

derive: Enables the derive macros for LdtkEntity and LdtkIntCell.

render: Enables rendering via bevy_ecs_tilemap’s render feature. Disable it if you want to run in headless mode.

atlas: Enables the atlas feature of bevy_ecs_tilemap. This is required for WASM support and also for tile spacing to work on Tile and AutoTile layers.

The derive and render features are enabled by default.

Re-exports

pub use crate::ldtk::EntityInstance;

Modules

app

Types and traits for hooking into the ldtk loading process via bevy::app::App.

ldtk

Contains all the types for serializing/deserializing an LDtk file.

prelude

use bevy_ecs_ldtk::prelude::*; to import commonly used items.

systems

System functions used by the plugin for processing ldtk files.

utils

Utility functions used internally by the plugin that have been exposed to the public api.

Structs

EntityIid

Component added to all LdtkEntitys by default.

GridCoords

Component that stores grid-based coordinate information.

IntGridCell

Component added to any IntGrid tile by default.

LayerMetadata

Component for storing some LDtk layer information on layer entities.

LdtkAsset

Main asset for loading ldtk files.

LdtkLevel

Secondary asset for loading ldtk files, specific to level data.

LdtkLevelLoader

LdtkLoader

LdtkPlugin

Adds the default systems, assets, and resources used by bevy_ecs_ldtk.

LdtkSettings

Settings resource for the plugin. Check out the documentation for each field type to learn more.

LdtkWorldBundle

Bundle for spawning LDtk worlds and their levels. The main bundle for using this plugin.

LevelSet

Component that determines the desired levels to be loaded for an LdtkWorldBundle.

ProcessLdtkApi

Schedule for processing this plugin’s ECS API, inserted after Update.

Respawn

Component that indicates that an LDtk level or world should respawn.

TileEnumTags

Component for storing user-defined, enum-based tags for a particular tile in an LDtk tileset definition.

TileMetadata

Component for storing user-defined custom data for a paticular tile in an LDtk tileset definition.

Worldly

Component that indicates that an ldtk entity should be a child of the world, not the level.

Enums

IntGridRendering

Option in LdtkSettings that determines the visual representation of IntGrid layers when they don’t have AutoTile rules.

LevelBackground

Option in LdtkSettings that dictates how the plugin handles level backgrounds.

LevelEvent

Events fired by the plugin related to level spawning/despawning.

LevelSelection

Resource for choosing which level(s) to spawn.

LevelSpawnBehavior

Option in LdtkSettings that determines level spawn behavior.

SetClearColor

Option in LdtkSettings that determines clear color behavior.

Type Definitions

LevelMap

Used in LdtkAsset. Key is the level iid.

TilesetMap

Used in LdtkAsset. Key is the tileset definition uid.

Derive Macros

LdtkEntity

LdtkIntCell