bevy_ecs_ldtk::assets
Trait LevelMetadataAccessor

pub trait LevelMetadataAccessor: RawLevelAccessor {
    // Required method
    fn get_level_metadata_by_iid(&self, iid: &String) -> Option<&LevelMetadata>;

    // Provided methods
    fn get_raw_level_by_iid(&self, iid: &String) -> Option<&Level> { ... }
    fn find_raw_level_by_level_selection(
        &self,
        level_selection: &LevelSelection,
    ) -> Option<&Level> { ... }
}

Convenience methods for types that store levels and level metadata.

Required Methods

fn get_level_metadata_by_iid(&self, iid: &String) -> Option<&LevelMetadata>

Returns a reference to the level metadata corresponding to the given level iid.

Provided Methods

fn get_raw_level_by_iid(&self, iid: &String) -> Option<&Level>

Immutable access to a level at the given level iid.

Note: all levels are considered raw.

fn find_raw_level_by_level_selection( &self, level_selection: &LevelSelection, ) -> Option<&Level>

Find the level matching the given LevelSelection.

This lookup is constant for LevelSelection::Iid and LevelSelection::Indices variants. The other variants require iterating through the levels to find the match.

Note: all levels are considered raw.