Module bevy_ecs_ldtk::systems

System functions used by the plugin for processing ldtk files.

Functions

apply_level_selection

Updates all LevelSet components according to the LevelSelection

apply_level_set

Triggers the spawning/despawning of levels according to LevelSet values.

clean_respawn_entities

Performs the “despawning” portion of the respawn process for Respawn entities.

detect_level_spawned_events

Returns the iids of levels that have spawned in this update.

fire_level_transformed_events

Fires LevelEvent::Transformed events for all the entities that spawned in the previous update.

process_ldtk_assets

Detects LdtkAsset events and spawns levels as children of the LdtkWorldBundle.

process_ldtk_levels

Performs all the spawning of levels, layers, chunks, bundles, entities, tiles, etc. when an LdtkLevelBundle is added.

worldly_adoption

Implements the functionality for Worldly components.